analog: rename getExAbsArchiveHoursMcEx to getArchiveHoursMcEx
