+++ src/analog/commands/downlink/getArchiveHoursMcEx.ts
