+++ src/analog/commands/uplink/getArchiveHoursMcEx.ts
