+++ src/analog/commands/uplink/index.ts
@@ -18,10 +18,10 @@ export * as getArchiveDaysMc from './getArchiveDaysMc.js';
 export * as getArchiveEvents from './getArchiveEvents.js';
 export * as getArchiveHours from './getArchiveHours.js';
 export * as getArchiveHoursMc from './getArchiveHoursMc.js';
+export * as getArchiveHoursMcEx from './getArchiveHoursMcEx.js';
 export * as getBatteryStatus from './getBatteryStatus.js';
 export * as getExAbsArchiveDaysMc from './getExAbsArchiveDaysMc.js';
 export * as getExAbsArchiveHoursMc from './getExAbsArchiveHoursMc.js';
-export * as getExAbsArchiveHoursMcEx from './getExAbsArchiveHoursMcEx.js';
 export * as getLmicInfo from './getLmicInfo.js';
 export * as getParameter from './getParameter.js';
 export * as hour from './hour.js';
