+++ src/analog/message/uplink.ts
@@ -58,10 +58,10 @@ toBytesMap[commands.getArchiveDaysMc.id] = commands.getArchiveDaysMc.toBytes;
 toBytesMap[commands.getArchiveEvents.id] = commands.getArchiveEvents.toBytes;
 toBytesMap[commands.getArchiveHours.id] = commands.getArchiveHours.toBytes;
 toBytesMap[commands.getArchiveHoursMc.id] = commands.getArchiveHoursMc.toBytes;
+toBytesMap[commands.getArchiveHoursMcEx.id] = commands.getArchiveHoursMcEx.toBytes;
 toBytesMap[commands.getBatteryStatus.id] = commands.getBatteryStatus.toBytes;
 toBytesMap[commands.getExAbsArchiveDaysMc.id] = commands.getExAbsArchiveDaysMc.toBytes;
 toBytesMap[commands.getExAbsArchiveHoursMc.id] = commands.getExAbsArchiveHoursMc.toBytes;
-toBytesMap[commands.getExAbsArchiveHoursMcEx.id] = commands.getExAbsArchiveHoursMcEx.toBytes;
 toBytesMap[commands.getLmicInfo.id] = commands.getLmicInfo.toBytes;
 toBytesMap[commands.getParameter.id] = commands.getParameter.toBytes;
 toBytesMap[commands.hour.id] = commands.hour.toBytes;
@@ -93,9 +93,9 @@ fromBytesMap[commands.getArchiveEvents.id] = commands.getArchiveEvents.fromBytes
 fromBytesMap[commands.getBatteryStatus.id] = commands.getBatteryStatus.fromBytes;
 fromBytesMap[commands.getArchiveHours.id] = commands.getArchiveHours.fromBytes;
 fromBytesMap[commands.getArchiveHoursMc.id] = commands.getArchiveHoursMc.fromBytes;
+fromBytesMap[commands.getArchiveHoursMcEx.id] = commands.getArchiveHoursMcEx.fromBytes;
 fromBytesMap[commands.getExAbsArchiveDaysMc.id] = commands.getExAbsArchiveDaysMc.fromBytes;
 fromBytesMap[commands.getExAbsArchiveHoursMc.id] = commands.getExAbsArchiveHoursMc.fromBytes;
-fromBytesMap[commands.getExAbsArchiveHoursMcEx.id] = commands.getExAbsArchiveHoursMcEx.fromBytes;
 fromBytesMap[commands.getLmicInfo.id] = commands.getLmicInfo.fromBytes;
 fromBytesMap[commands.getParameter.id] = commands.getParameter.fromBytes;
 fromBytesMap[commands.hour.id] = commands.hour.fromBytes;
@@ -126,9 +126,9 @@ nameMap[commands.getArchiveEvents.id] = commands.getArchiveEvents.name;
 nameMap[commands.getBatteryStatus.id] = commands.getBatteryStatus.name;
 nameMap[commands.getArchiveHours.id] = commands.getArchiveHours.name;
 nameMap[commands.getArchiveHoursMc.id] = commands.getArchiveHoursMc.name;
+nameMap[commands.getArchiveHoursMcEx.id] = commands.getArchiveHoursMcEx.name;
 nameMap[commands.getExAbsArchiveDaysMc.id] = commands.getExAbsArchiveDaysMc.name;
 nameMap[commands.getExAbsArchiveHoursMc.id] = commands.getExAbsArchiveHoursMc.name;
-nameMap[commands.getExAbsArchiveHoursMcEx.id] = commands.getExAbsArchiveHoursMcEx.name;
 nameMap[commands.getLmicInfo.id] = commands.getLmicInfo.name;
 nameMap[commands.getParameter.id] = commands.getParameter.name;
 nameMap[commands.hour.id] = commands.hour.name;
